Experimental/Sinhala-transcription-OpenAI/requirements.txt
openai>=1.3.5
streamlit
pydub
ffmpeg-python
python-dotenv